v6.2.0 patch - Add m4 files for curl, jansson and microhhtpd. - EAROwner added. - EARD security fixes. - Improve file permission and ownership. - `erun` command sets JobID to 1 by default. - Default CFLAGS changed to add `-Wall -Werror -O2 -g` to improve code quality. - Lighter CI pipeline. - Prevent calling `mgt\_cpufreq\_governor\_available` when the symbol is NULL. - Fix multiple opened fd checks that might be zero. - Fixed setuid security issues in ereport, eacct, ear-info and econtrol. - EARD rAPI: Generic message is not propagating properly. - IPMI sensors. - Fixed "Policies default PState is overwritten in EARD's main function."

+++ src/management/gpu/archs/mgt_pvc_power_hwmon.c
@@ -59,7 +59,7 @@ static void init_pv_freqs()
         exit(0);
     }
     for (uint dv = 0; dv < pvc_gpus_num_devices; dv++) {
-        gpu_clock_list[dv] = calloc(1, sizeof(uint));
+        gpu_clock_list[dv] = calloc(1, sizeof(ulong));
         if (gpu_clock_list[dv] == NULL) {
             verbose(0, "Error not enough memory");
             exit(0);
